Refactor to use vars middleware
